change LunNumber from float64 to int

+++ remote/mounter/initiator/linuxfc.go
@@ -53,11 +53,10 @@ func (lfc *linuxFibreChannel) hasFCSupport() bool {
 // may be '-' wildcards if unable to determine them.
 func (lfc *linuxFibreChannel) getHBAChannelScsiTarget(volumeMountProperties *resources.VolumeMountProperties) string {
 	//TODO: get channel and target
-	if volumeMountProperties.LunNumber == float64(-1) {
+	if volumeMountProperties.LunNumber == -1 {
 		return "- - -"
 	}
-	// use %g to print a float64 to int
-	return fmt.Sprintf("- - %g", volumeMountProperties.LunNumber)
+	return fmt.Sprintf("- - %d", volumeMountProperties.LunNumber)
 }
 
 // GetHBAs return all the FC HBAs in the system
